Tentacle AI now works
When activated, selects a random amount and then random tentacles of said amount and 'attacks' with them.
Deploy time is how long theyre out for, Wait time is the time between 'attacks'. The longer an attack goes without happening the more likely it is to happen as well.
(1hr30) my dumbass was accidently duping the script each time and couldnt figure out why